+++ packages/gatsby-plugin-styled-components/CHANGELOG.md
@@ -3,8 +3,8 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-<a name="3.0.0-beta.1"></a>
-# [3.0.0-beta.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-styled-components/compare/gatsby-plugin-styled-components@3.0.0-beta.0...gatsby-plugin-styled-components@3.0.0-beta.1) (2018-06-17)
+<a name="3.0.0-beta.2"></a>
+# [3.0.0-beta.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-styled-components/compare/gatsby-plugin-styled-components@3.0.0-beta.1...gatsby-plugin-styled-components@3.0.0-beta.2) (2018-06-20)
 
 **Note:** Version bump only for package gatsby-plugin-styled-components
 
@@ -12,7 +12,14 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+<a name="3.0.0-beta.1"></a>
+
+# [3.0.0-beta.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-styled-components/compare/gatsby-plugin-styled-components@3.0.0-beta.0...gatsby-plugin-styled-components@3.0.0-beta.1) (2018-06-17)
+
+**Note:** Version bump only for package gatsby-plugin-styled-components
+
 <a name="3.0.0-beta.0"></a>
+
 # [3.0.0-beta.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-styled-components/compare/gatsby-plugin-styled-components@2.0.11...gatsby-plugin-styled-components@3.0.0-beta.0) (2018-06-17)
 
 **Note:** Version bump only for package gatsby-plugin-styled-components