+++ packages/gatsby-source-filesystem/CHANGELOG.md
@@ -3,8 +3,8 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-<a name="2.0.1-beta.1"></a>
-## [2.0.1-beta.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-filesystem/compare/gatsby-source-filesystem@2.0.1-beta.0...gatsby-source-filesystem@2.0.1-beta.1) (2018-06-17)
+<a name="2.0.1-beta.3"></a>
+## [2.0.1-beta.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-filesystem/compare/gatsby-source-filesystem@2.0.1-beta.2...gatsby-source-filesystem@2.0.1-beta.3) (2018-06-21)
 
 **Note:** Version bump only for package gatsby-source-filesystem
 
@@ -12,7 +12,23 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+<a name="2.0.1-beta.2"></a>
+## [2.0.1-beta.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-filesystem/compare/gatsby-source-filesystem@2.0.1-beta.1...gatsby-source-filesystem@2.0.1-beta.2) (2018-06-20)
+
+**Note:** Version bump only for package gatsby-source-filesystem
+
+
+
+
+
+<a name="2.0.1-beta.1"></a>
+
+## [2.0.1-beta.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-filesystem/compare/gatsby-source-filesystem@2.0.1-beta.0...gatsby-source-filesystem@2.0.1-beta.1) (2018-06-17)
+
+**Note:** Version bump only for package gatsby-source-filesystem
+
 <a name="2.0.1-beta.0"></a>
+
 ## [2.0.1-beta.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-filesystem/compare/gatsby-source-filesystem@1.5.39...gatsby-source-filesystem@2.0.1-beta.0) (2018-06-17)
 
 **Note:** Version bump only for package gatsby-source-filesystem